requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
azure-cli>=2.53.0
azure-mgmt-compute>=30.0.0
azure-identity>=1.14.0
aiohttp>=3.9.0
numpy>=1.24.0
psutil>=5.9.0
httptools>=0.6.0
uvloop>=0.19.0
python-jose>=3.3.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.25.0
flash-attn>=2.3.0
einops>=0.7.0
ninja>=1.11.0

